共通のURL判定関数を追加
