+++ app/client/src/utils/url.ts
@@ -1,9 +1,1 @@
-export function isUrl(value?: string): boolean {
-  if (!value) return false;
-  try {
-    const url = new URL(value.trim());
-    return url.protocol === "http:" || url.protocol === "https:";
-  } catch {
-    return false;
-  }
-}
+export { isUrl } from "../../../shared/url.ts";
