+++ app/shared/url.ts
@@ -1,0 +1,9 @@
+export function isUrl(value?: string): boolean {
+  if (!value) return false;
+  try {
+    const url = new URL(value.trim());
+    return url.protocol === "http:" || url.protocol === "https:";
+  } catch {
+    return false;
+  }
+}
